etc: processTag타입 수정
