--- types/lecture.type.ts
@@ -1,7 +1,0 @@
-export interface IProcessData {
-  category: string;
-  skill: string;
-  tags: string[];
-  timeAsc: boolean | null;
-  priceAsc: boolean | null;
-}
