+++ types/lectures.type.ts
@@ -33,3 +33,18 @@ export interface LecturesResultAllData {
   categoryId: number;
   skillId: number;
 }
+
+export interface IProcessData {
+  category: string;
+  skill: string;
+  tags: string[];
+  timeAsc: boolean | null;
+  priceAsc: boolean | null;
+}
+export interface ProcessDataState {
+  category: string;
+  skill: string;
+  tags: string[];
+  timeAsc: boolean | null | undefined;
+  priceAsc: boolean | null | undefined;
+}
